requirements.txt
exceptiongroup==1.2.0
iniconfig==2.0.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pdf2image==1.17.0
pillow==10.3.0
pluggy==1.4.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pytest==8.1.1
python-barcode==0.15.1
pyzbar==0.1.9
tomli==2.0.1
